+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+WIP
+
+## [0.1.13] - 2023-11-11
+
 ### Added
 
 - Status line to show loading while communicating with database.
@@ -134,7 +138,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Execution results view supporting only successful fetched results.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/strowk/tisq/compare/v0.1.12...HEAD
+[Unreleased]: https://github.com/strowk/tisq/compare/v0.1.13...HEAD
+[0.1.13]: https://github.com/strowk/tisq/compare/v0.1.12...v0.1.13
 [0.1.12]: https://github.com/strowk/tisq/compare/v0.1.11...v0.1.12
 [0.1.11]: https://github.com/strowk/tisq/compare/v0.1.10...v0.1.11
 [0.1.10]: https://github.com/strowk/tisq/compare/v0.1.9...v0.1.10